pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentos-platform"
version = "0.3.1"
description = "The Operating System for AI Agents — Build, Test, Deploy, Monitor, and Govern."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    {name = "Suketh Reddy Produtoor", email = "produtoor.s@northeastern.edu"},
]
keywords = ["ai", "agents", "llm", "openai", "anthropic", "governance", "monitoring"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.30.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "python-multipart>=0.0.6",
    "typer>=0.9.0",
    "kubernetes-asyncio>=28.0.0",
    "pyyaml>=6.0",
    "cachetools>=5.0.0",
    "networkx>=3.0",
    "rank_bm25>=0.2.2",
    "sentence-transformers>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
redis = [
    "redis>=5.0.0",
]
rag = [
    "chromadb>=0.4.0",
    "pinecone-client>=2.2.0",
    "pgvector>=0.2.0",
    "psycopg[binary]>=3.0",
]

[project.urls]
Homepage = "https://github.com/sukethrp/agentos"
Repository = "https://github.com/sukethrp/agentos"
Issues = "https://github.com/sukethrp/agentos/issues"

[project.scripts]
agentos = "agentos.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff.lint.per-file-ignores]
# Deferred/lazy imports for optional features
"src/agentos/web/app.py" = ["E402"]
# conftest: sys.path setup before agentos import
"tests/conftest.py" = ["E402"]